+++ src/test/java/org/folio/search/service/reindex/jdbc/CallNumberRepositoryIT.java
@@ -103,9 +103,11 @@ class CallNumberRepositoryIT {
       .extracting("callNumber", "instances")
       .contains(
         tuple("number1",
-          List.of(mapOf("count", 1, "locationId", null, "shared", false, "tenantId", TENANT_ID, "typeId", null))),
+          List.of(mapOf("count", 0, "instanceId", List.of("9f8febd1-e96c-46c4-a5f4-84a45cc499a2"),
+            "instanceTitle", null, "locationId", null, "shared", false, "tenantId", TENANT_ID, "typeId", null))),
         tuple("number2",
-          List.of(mapOf("count", 1, "locationId", null, "shared", false, "tenantId", TENANT_ID, "typeId", null))));
+          List.of(mapOf("count", 0, "instanceId", List.of("9f8febd1-e96c-46c4-a5f4-84a45cc499a2"),
+            "instanceTitle", null, "locationId", null, "shared", false, "tenantId", TENANT_ID, "typeId", null))));
   }
 
   private static List<UUID> getList(int size) {